fe/generator/pyproject.toml
[project]
name = "uxvmarket-generator"
version = "0.1.0"
description = "Reads canonical vendor JSON and emits Astro content files for fe/site"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[tool.uv]
package = false
